feat(mcp): drain EA/ontology tool domain into a scoped pack (EP-8DC217EB BET-4)
Move the eight Enterprise-Architecture / ArchiMate ontology-graph tools
(create_ea_element, create_ea_relationship, classify_ea_element,
query_ontology_graph, run_traversal_pattern, describe_ea_view,
import_archimate, export_archimate) out of the mcp-tools.ts executeTool
switch into apps/web/lib/mcp/packs/ea-ontology-pack.ts. Definitions and
handler bodies are relocated verbatim; grants mirror agent-grants.ts
TOOL_TO_GRANTS. Registered via pack-registry.ts (one import + one array
entry); mcp-tools.ts only loses the drained defs and case blocks. Ratchet
baseline shrunk 202 -> 194.

Process-Spine-Decision: EP-8DC217EB plan §4 BET-4 — drain the EA/ontology tool domain into a scoped pack (registered via pack-registry.ts); ratchet baseline shrunk; behavior-preserving relocation.
