+++ apps/web/lib/mcp/pack-registry.ts
@@ -45,6 +45,7 @@ import { nonprodLeasePack } from "./packs/nonprod-lease-pack";
 import { knowledgePack } from "./packs/knowledge-pack";
 import { demandScoringPack } from "./packs/demand-scoring-pack";
 import { workforcePack } from "./packs/workforce-pack";
+import { eaOntologyPack } from "./packs/ea-ontology-pack";
 
 export const TOOL_PACK_REGISTRY = composeToolPacks([
   deliberationSiemPack,
@@ -76,4 +77,5 @@ export const TOOL_PACK_REGISTRY = composeToolPacks([
   knowledgePack,
   demandScoringPack,
   workforcePack,
+  eaOntologyPack,
 ]);
