+++ scripts/mcp-tool-pack-baseline.json
@@ -22,7 +22,6 @@
   "cancel_thread",
   "check_deployment_windows",
   "check_sandbox",
-  "classify_ea_element",
   "compare_versions",
   "configure_gateway_scan",
   "confirm_taxonomy_placement",
@@ -31,8 +30,6 @@
   "create_build_epic",
   "create_customer_account",
   "create_digital_product",
-  "create_ea_element",
-  "create_ea_relationship",
   "create_epic",
   "create_licensing_readiness_issue",
   "create_marketing_asset_task",
@@ -44,7 +41,6 @@
   "create_release_bundle",
   "create_scheduled_agent_task",
   "deploy_feature",
-  "describe_ea_view",
   "describe_model",
   "diagnose_sandbox",
   "discovery_sweep",
@@ -57,7 +53,6 @@
   "evaluate_tool",
   "execute_promotion",
   "explain_blast_radius",
-  "export_archimate",
   "extract_brand_design_system",
   "fetch_public_website",
   "find_related_tests",
@@ -84,7 +79,6 @@
   "get_thread_result",
   "grok_signin_start",
   "grok_signin_status",
-  "import_archimate",
   "inspect_build_code_impact",
   "iterate_sandbox",
   "link_backlog_item_to_epic",
@@ -117,7 +111,6 @@
   "publish_to_linkedin",
   "publish_wiki_overlay_pages",
   "query_backlog",
-  "query_ontology_graph",
   "query_version_history",
   "read_codebase_manifest",
   "read_project_file",
@@ -147,7 +140,6 @@
   "run_sandbox_command",
   "run_sandbox_tests",
   "run_tool_script",
-  "run_traversal_pattern",
   "run_ux_test",
   "saveBuildEvidence",
   "save_build_notes",
